💄全体のレイアウトを調整する
